linuganex_api/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.5
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
kombu==5.2.4
outcome==1.2.0
prompt-toolkit==3.0.36
pycparser==2.21
PySocks==1.7.1
pytz==2022.7
redis==4.4.1
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wsproto==1.2.0
